// src/app/interfaces/LaunchesClass.ts
export interface LaunchesClass{
    flight_number:           number;
    mission_name:            string;
    mission_id:              string[];
    upcoming:                boolean;
    launch_year:             string;
    launch_date_unix:        number;
    launch_date_utc:         string;
    launch_date_local:       string;
    is_tentative:            boolean;
    tentative_max_precision: TentativeMaxPrecision;
    tbd:                     boolean;
    launch_window:           number | null;
    rocket:                  Rocket;
    ships:                   string[];
    telemetry:               Telemetry;
    launch_site:             LaunchSite;
    launch_success:          boolean | null;
    launch_failure_details?: LaunchFailureDetails;
    links:                   Links;
    details:                 null | string;
    static_fire_date_utc:    null | string;
    static_fire_date_unix:   number | null;
    timeline:                { [key: string]: number | null } | null;
    crew:                    any[] | null;
    last_date_update?:       string;
    last_ll_launch_date?:    null | string;
    last_ll_update?:         null | string;
    last_wiki_launch_date?:  string;
    last_wiki_revision?:     string;
    last_wiki_update?:       string;
    launch_date_source?:     LaunchDateSource;
}

export enum LaunchDateSource {
    LaunchLibrary = "launch_library",
    Wiki = "wiki",
}

export interface LaunchFailureDetails {
    time:     number;
    altitude: number | null;
    reason:   string;
}

export interface LaunchSite {
    site_id:        SiteID;
    site_name:      SiteName;
    site_name_long: SiteNameLong;
}

export enum SiteID {
    CcafsSlc40 = "ccafs_slc_40",
    KscLc39A = "ksc_lc_39a",
    KwajaleinAtoll = "kwajalein_atoll",
    VafbSlc4E = "vafb_slc_4e",
}

export enum SiteName {
    CcafsSlc40 = "CCAFS SLC 40",
    KscLc39A = "KSC LC 39A",
    KwajaleinAtoll = "Kwajalein Atoll",
    VafbSlc4E = "VAFB SLC 4E",
}

export enum SiteNameLong {
    CapeCanaveralAirForceStationSpaceLaunchComplex40 = "Cape Canaveral Air Force Station Space Launch Complex 40",
    KennedySpaceCenterHistoricLaunchComplex39A = "Kennedy Space Center Historic Launch Complex 39A",
    KwajaleinAtollOmelekIsland = "Kwajalein Atoll Omelek Island",
    VandenbergAirForceBaseSpaceLaunchComplex4E = "Vandenberg Air Force Base Space Launch Complex 4E",
}

export interface Links {
    mission_patch:       null | string;
    mission_patch_small: null | string;
    reddit_campaign:     null | string;
    reddit_launch:       null | string;
    reddit_recovery:     null | string;
    reddit_media:        null | string;
    presskit:            null | string;
    article_link:        null | string;
    wikipedia:           null | string;
    video_link:          null | string;
    youtube_id:          null | string;
    flickr_images:       string[];
}

export interface Rocket {
    rocket_id:    RocketID;
    rocket_name:  RocketName;
    rocket_type:  RocketType;
    first_stage:  FirstStage;
    second_stage: SecondStage;
    fairings:     Fairings | null;
}

export interface Fairings {
    reused:           boolean | null;
    recovery_attempt: boolean | null;
    recovered:        boolean | null;
    ship:             Ship | null;
}

export enum Ship {
    Gomschief = "GOMSCHIEF",
    Gomstree = "GOMSTREE",
    Gosearcher = "GOSEARCHER",
}

export interface FirstStage {
    cores: Core[];
}
// ...
export enum RocketID {
    Falcon1 = "falcon1",
    Falcon9 = "falcon9",
    Falconheavy = "falconheavy",
}

export enum RocketName {
    Falcon1 = "Falcon 1",
    Falcon9 = "Falcon 9",
    FalconHeavy = "Falcon Heavy",
}

export enum RocketType {
    Ft = "FT",
    MerlinA = "Merlin A",
    MerlinC = "Merlin C",
    V10 = "v1.0",
    V11 = "v1.1",
}

export interface SecondStage {
    block:    number | null;
    payloads: Payload[];
}
// ...
export interface Telemetry {
    flight_club: null | string;
}

export enum TentativeMaxPrecision {
    Hour = "hour",
}