// src/app/interfaces/FinalDataClass.ts
// Generated by https://quicktype.io

export interface FinalDataClass {
    flight_number:    number;
    mission_name:     string;
    launch_date_unix: number;
    details:          null | string;
    rocket:           Rocket;
}

export interface Rocket {
    rocket_id:       string;
    rocket_name:     string;
    active:          boolean;
    cost_per_launch: number;
    company:         string;
}